# Conflicts:
#	app/src/main/java/com/example/spacex_app/data/network/retrofit/service/InfoService.kt
#	app/src/main/java/com/example/spacex_app/data/repository/InfoRepository.kt
#	app/src/main/java/com/example/spacex_app/di/ModuleService.kt
#	app/src/main/java/com/example/spacex_app/presentation/viewmodel/CompanyViewModel.kt

+++ app/src/main/java/com/example/spacex_app/data/network/response/rocketResponse/EngineResponse.kt
@@ -11,5 +11,5 @@ data class EngineResponse(
     var propellant_2: String? = "",
     var thrust_sea_level: ThrustSeaLevelResponse? = ThrustSeaLevelResponse(),
     var thrust_vacuum: ThrustVacuumResponse? = ThrustVacuumResponse(),
-    var thrust_to_weight: Int? = 0
+    var thrust_to_weight: Double? = 0.0
 )
